scenarios/hpc_ai_lab/src/requirements.txt
# data science
numpy
scipy
pandas>=1.2.0
adlfs>=2021.8.1
scikit-learn
torch
torchvision

# tracking
mlflow
azureml-mlflow

# storage
azure.storage.blob